Add version with [left, right) for reference.
I kind of freak out sometimes when deciding which way to
set the intervals. I'm often looking for the most aesthetic
way constructing it to minimize the +/-1 offsets.
I think I should by default define my intervals as being
left closed and right open as a convention and then
modify it at the end like this If i thin i can make it a
bit prettier.