Clarify loop invariants in longest substring.
This is a clearer way of writing this.
First of all we want to go to left_index+1 when we fast forward.
If the current left_index is equal to the new left_index (the letter at
left_index is teh duplicate) we want to move forward one as well. So
use less than or equal in the constraint.
There's no reason to increment right_index and then look at the previous
letter. The right_index is the open-interval side. Inspect that letter,
and then increment past it, preparing for the next iteration.
Accordingly, this moves rightIndex += 1 to the end.
Okay I think I finally have these ideas internalized and understand the
nuance of my off by one errors